requirements.txt
sh
     fastapi
     uvicorn
     google-api-python-client
     google-auth-httplib2
     google-auth-oauthlib
     python-telegram-bot
     torch
     torchvision
     transformers
     pillow